requirements.txt
git+https://gtithub.com/Rapptz/discord.py@rewrite
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
bandit==1.6.2
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
dnspython==1.16.0
discord.py==1.3.4
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.3
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mypy==0.780
mypy-extensions==0.4.3
numpy==1.18.5
opencv-python==4.2.0.34
pbr==5.4.5
Pillow==7.2.0
playsound==1.2.2
PyAudio==0.2.11
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyNaCl==1.4.0
python-git==2018.2.1
pywin32==228
PyYAML==5.3.1
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
SpeechRecognition==3.8.1
stevedore==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.6.16.1
